🚚 move Qiskit code to `plugin` directory
This is mostly to avoid a naming conflict on editable installs because the `src/mqt/qmap` directory contained a subdirectory `qiskit`.
