+++ src/mqt/qmap/load_calibration.py
@@ -24,11 +24,11 @@ def load_calibration(architecture: Architecture, calibration: str | Target | Bac
     if isinstance(calibration, str):
         architecture.load_properties(calibration)
     elif isinstance(calibration, BackendProperties):
-        from mqt.qmap.qiskit.backend import import_backend_properties
+        from mqt.qmap.plugins.qiskit import import_backend_properties
 
         architecture.load_properties(import_backend_properties(calibration))
     elif isinstance(calibration, Target):
-        from mqt.qmap.qiskit.backend import import_target
+        from mqt.qmap.plugins.qiskit import import_target
 
         architecture.load_properties(import_target(calibration))
     else:  # pragma: no cover
